concepts/os/requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2
google-auth==2.17.3
google-auth-oauthlib==1.0.0
gspread==5.8.0
idna==3.4
image==1.5.33
numpy==1.24.2
oauthlib==3.2.2
pandas==2.0.0
Pillow==9.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-dateutil==2.8.2
python-resize-image==1.1.20
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
